Fix up namespaces (plain renames)
