+++ choreographer/browsers/README.txt
