+++ choreographer/browsers/__init__.py
@@ -1,3 +1,5 @@
+"""browsers contains implementations of browsers that choreographer can open."""
+
 from ._errors import BrowserClosedError, BrowserFailedError
 from .chromium import Chromium
