+++ choreographer/channels/__init__.py
@@ -1,3 +1,5 @@
+"""Channels are classes that choreo and the browser use to communicate."""
+
 from ._errors import BlockWarning, ChannelClosedError, JSONError
 from .pipe import Pipe
