+++ choreographer/cli/__init__.py
@@ -1,10 +1,12 @@
-from ._cli_utils import (
+"""cli provides some tools that are used on the commandline (and to download chrome)."""
+
+from ._cli_utils_no_qa import diagnose
+from .cli import (
     get_chrome,
     get_chrome_cli,
     get_chrome_download_path,
     get_chrome_sync,
 )
-from ._cli_utils_no_qa import diagnose
 
 __all__ = [
     "diagnose",
