+++ choreographer/protocol/__init__.py
@@ -1,0 +1,5 @@
+"""Subpackage protocol provides various implementations of Session and Target."""
+
+from . import sync
+
+__all__ = ["sync"]
