+++ choreographer/utils/__init__.py
@@ -1,11 +1,10 @@
-from ._kill import kill
+"""Utils contains functions and class that primarily help us with the OS."""
+
 from ._tmpfile import TmpDirectory, TmpDirWarning
-from ._which import browser_which, get_browser_path
+from ._which import get_browser_path
 
 __all__ = [
     "TmpDirWarning",
     "TmpDirectory",
-    "browser_which",
     "get_browser_path",
-    "kill",
 ]
